CMUpdater: Check permissions for external storage
For operations that involve writing to external storage, ensure
permissions are checked first. Notably, this fixes deletion operations
if data for the CMUpdater app is cleared -- there was previously a
null pointer exception on delete-all when permissions were not already
confirmed.

This commit leaves in-tact the existing behavior of not listing
existing updates on application start if permissions to external
storage have not been granted.

+++ src/com/cyanogenmod/updater/UpdatesSettings.java
@@ -84,6 +84,10 @@ public class UpdatesSettings extends PreferenceActivity implements
 
     private static final int PERMISSIONS_REQUEST_WRITE_EXTERNAL_STORAGE = 1;
 
+    private static final int PERMISSIONS_REQUEST_FOR_DOWNLOAD = 0;
+    private static final int PERMISSIONS_REQUEST_FOR_DELETE_ZIP = 1;
+    private static final int PERMISSIONS_REQUEST_FOR_DELETE_ALL = 2;
+
     private SharedPreferences mPrefs;
     private ListPreference mUpdateCheck;
 
@@ -98,7 +102,11 @@ public class UpdatesSettings extends PreferenceActivity implements
     private DownloadManager mDownloadManager;
     private boolean mDownloading = false;
     private long mDownloadId;
-    private String mFileName;
+    private String mDownloadFileName;
+
+    private String mDeleteFileName = "";
+
+    private int mPermissionsReqType;
 
     private Handler mUpdateHandler = new Handler();
 
@@ -278,14 +286,14 @@ public class UpdatesSettings extends PreferenceActivity implements
                 if (status == DownloadManager.STATUS_PENDING
                         || status == DownloadManager.STATUS_RUNNING
                         || status == DownloadManager.STATUS_PAUSED) {
-                    mFileName = uri.getLastPathSegment();
+                    mDownloadFileName = uri.getLastPathSegment();
                 }
             }
             if (c != null) {
                 c.close();
             }
         }
-        if (mDownloadId < 0 || mFileName == null) {
+        if (mDownloadId < 0 || mDownloadFileName == null) {
             resetDownloadState();
         }
 
@@ -325,20 +333,8 @@ public class UpdatesSettings extends PreferenceActivity implements
 
         // We have a match, get ready to trigger the download
         mDownloadingPreference = pref;
-
-        // But check permissions first - download will be started in the callback
-        int permissionCheck = ContextCompat.checkSelfPermission(pref.getContext(),
-                android.Manifest.permission.WRITE_EXTERNAL_STORAGE);
-        if (permissionCheck == PackageManager.PERMISSION_GRANTED) {
-            // permission already granted, start the download
-            startDownload();
-        } else {
-            // permission not granted, request it from the user
-            ActivityCompat.requestPermissions(this,
-                    new String[]{android.Manifest.permission.WRITE_EXTERNAL_STORAGE},
-                    PERMISSIONS_REQUEST_WRITE_EXTERNAL_STORAGE);
-        }
-
+        mPermissionsReqType = PERMISSIONS_REQUEST_FOR_DOWNLOAD;
+        requestStoragePermission(pref.getContext());
     }
 
     private Runnable mUpdateProgress = new Runnable() {
@@ -410,7 +406,7 @@ public class UpdatesSettings extends PreferenceActivity implements
 
     @Override
     public void onStopDownload(final UpdatePreference pref) {
-        if (!mDownloading || mFileName == null || mDownloadId < 0) {
+        if (!mDownloading || mDownloadFileName == null || mDownloadId < 0) {
             pref.setStyle(UpdatePreference.STYLE_NEW);
             resetDownloadState();
             return;
@@ -489,7 +485,7 @@ public class UpdatesSettings extends PreferenceActivity implements
 
     private void resetDownloadState() {
         mDownloadId = -1;
-        mFileName = null;
+        mDownloadFileName = null;
         mDownloading = false;
         mDownloadingPreference = null;
     }
@@ -645,7 +641,7 @@ public class UpdatesSettings extends PreferenceActivity implements
             }
 
             // Determine the preference style and create the preference
-            boolean isDownloading = ui.getFileName().equals(mFileName);
+            boolean isDownloading = ui.getFileName().equals(mDownloadFileName);
             int style;
 
             if (isDownloading) {
@@ -691,22 +687,31 @@ public class UpdatesSettings extends PreferenceActivity implements
         }
     }
 
+    private void deleteZipUpdate() {
+        if (mDeleteFileName.isEmpty()) {
+            Log.e(TAG, "deleteZipUpdate: File name not specified");
+            return;
+        }
+
+        File zipFileToDelete = new File(mUpdateFolder, mDeleteFileName);
+
+        if (zipFileToDelete.exists()) {
+            zipFileToDelete.delete();
+        } else {
+            Log.d(TAG, "Update to delete not found");
+            return;
+        }
+
+        String message = getString(R.string.delete_single_update_success_message, mDeleteFileName);
+        Toast.makeText(this, message, Toast.LENGTH_SHORT).show();
+    }
+
     @Override
     public void onDeleteUpdate(UpdatePreference pref) {
-        final String fileName = pref.getKey();
-
         if (mUpdateFolder.exists() && mUpdateFolder.isDirectory()) {
-            File zipFileToDelete = new File(mUpdateFolder, fileName);
-
-            if (zipFileToDelete.exists()) {
-                zipFileToDelete.delete();
-            } else {
-                Log.d(TAG, "Update to delete not found");
-                return;
-            }
-
-            String message = getString(R.string.delete_single_update_success_message, fileName);
-            Toast.makeText(this, message, Toast.LENGTH_SHORT).show();
+            mDeleteFileName = pref.getKey();
+            mPermissionsReqType = PERMISSIONS_REQUEST_FOR_DELETE_ZIP;
+            requestStoragePermission(pref.getContext());
         } else if (!mUpdateFolder.exists()) {
             Toast.makeText(this, R.string.delete_updates_noFolder_message, Toast.LENGTH_SHORT).show();
         } else {
@@ -726,10 +731,35 @@ public class UpdatesSettings extends PreferenceActivity implements
                 if (grantResults.length > 0
                         && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
                     // permission was granted, yay!
-                    startDownload();
+                    switch (mPermissionsReqType) {
+                        case PERMISSIONS_REQUEST_FOR_DOWNLOAD:
+                            startDownload();
+                            break;
+                        case PERMISSIONS_REQUEST_FOR_DELETE_ZIP:
+                            deleteZipUpdate();
+                            break;
+                        case PERMISSIONS_REQUEST_FOR_DELETE_ALL:
+                            deleteOldUpdates();
+                            break;
+                        default:
+                            break;
+                    }
                 } else {
-                    // permission was not granted, cannot download
-                    mDownloadingPreference = null;
+                    // permission was not granted
+                    switch (mPermissionsReqType) {
+                        case PERMISSIONS_REQUEST_FOR_DOWNLOAD:
+                            mDownloadingPreference = null;
+                            break;
+                        case PERMISSIONS_REQUEST_FOR_DELETE_ZIP:
+                            mDeleteFileName = "";
+                            Log.e(TAG, "deleteZipUpdate: Permission not granted");
+                            break;
+                        case PERMISSIONS_REQUEST_FOR_DELETE_ALL:
+                            Log.e(TAG, "deleteOldUpdates: Permission not granted");
+                            break;
+                        default:
+                            break;
+                    }
                     new AlertDialog.Builder(this)
                             .setTitle(R.string.permission_not_granted_dialog_title)
                             .setMessage(R.string.permission_not_granted_dialog_message)
@@ -758,7 +788,7 @@ public class UpdatesSettings extends PreferenceActivity implements
         ImageView updatesButton = mDownloadingPreference.getUpdatesButton();
         updatesButton.setEnabled(false);
 
-        mFileName = ui.getFileName();
+        mDownloadFileName = ui.getFileName();
         mDownloading = true;
 
         // Start the download
@@ -778,12 +808,38 @@ public class UpdatesSettings extends PreferenceActivity implements
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
                         // We are OK to delete, trigger it
-                        deleteOldUpdates();
-                        updateLayout();
+                        mPermissionsReqType = PERMISSIONS_REQUEST_FOR_DELETE_ALL;
+                        requestStoragePermission(getApplicationContext());
                     }
                 })
                 .setNegativeButton(R.string.dialog_cancel, null)
                 .show();
+    }
+
+    private void requestStoragePermission(Context context) {
+            int permissionCheck = ContextCompat.checkSelfPermission(context,
+                    android.Manifest.permission.WRITE_EXTERNAL_STORAGE);
+            if (permissionCheck == PackageManager.PERMISSION_GRANTED) {
+                // permission already granted, go ahead
+                switch (mPermissionsReqType) {
+                    case PERMISSIONS_REQUEST_FOR_DOWNLOAD:
+                        startDownload();
+                        break;
+                    case PERMISSIONS_REQUEST_FOR_DELETE_ZIP:
+                        deleteZipUpdate();
+                        break;
+                    case PERMISSIONS_REQUEST_FOR_DELETE_ALL:
+                        deleteOldUpdates();
+                        break;
+                    default:
+                        break;
+                }
+            } else {
+                // permission not granted, request it from the user
+                ActivityCompat.requestPermissions(this,
+                        new String[]{android.Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                        PERMISSIONS_REQUEST_WRITE_EXTERNAL_STORAGE);
+            }
     }
 
     private boolean deleteOldUpdates() {
@@ -801,12 +857,18 @@ public class UpdatesSettings extends PreferenceActivity implements
             success = false;
             Toast.makeText(this, R.string.delete_updates_failure_message, Toast.LENGTH_SHORT).show();
         }
+        updateLayout();
         return success;
     }
 
     private static boolean deleteDir(File dir) {
         if (dir.isDirectory()) {
             String[] children = dir.list();
+            if (children == null) {
+                // An IO or permissions failure; don't crash
+                Log.e(TAG, "deleteDir: dir.list() failed, check storage permissions");
+                return false;
+            }
             for (String aChildren : children) {
                 boolean success = deleteDir(new File(dir, aChildren));
                 if (!success) {
